+++ CHANGES.md
@@ -2,8 +2,9 @@
 
 ## Unreleased
 * Built using Avro 1.10.2
-* Built using Gradle 6.8.3
+* Built using Gradle 7.0-rc-1
 * Updated compatibility testing through Gradle 6.8.3
+* Updated compatibility testing through Kotlin 1.4.32
 
 ## 1.0.0
 * Published to Maven Central (no longer published to JCenter)